pyproject.toml
[project]
name = "quantum-gan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "cirq>=1.5.0",
    "ipykernel>=6.29.5",
    "keras>=3.10.0",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "scikit-learn>=1.7.0",
    "tensorflow>=2.19.0",
    "torch>=2.7.1",
]
